pyproject.toml
[tool.poetry]
name = "web-scrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.1"
request = "^2020.7.1"
pandas = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
